+++ Engine1/Direct3DUtil.cpp
@@ -5,7 +5,7 @@
 using namespace Engine1;
 
 void Direct3DUtil::setResourceName( ID3D11DeviceChild& child, const std::string& name ) {
-	child.SetPrivateData( WKPDID_D3DDebugObjectName, name.size(), name.c_str() );
+	child.SetPrivateData( WKPDID_D3DDebugObjectName, (unsigned int)name.size(), name.c_str() );
 }
 
 int Direct3DUtil::getRefCount( IUnknown& object )
